split cache tests
