+++ src/NServiceBus.Persistence.AzureStorage.ComponentTests/Subscriptions/TestMessage.cs
@@ -1,0 +1,6 @@
+﻿namespace NServiceBus.Persistence.AzureStorage.ComponentTests.Subscriptions
+{
+    class TestMessage
+    {
+    }
+}
