+++ src/NServiceBus.Persistence.AzureStorage.ComponentTests/Subscriptions/TestMessagea.cs
@@ -1,0 +1,6 @@
+﻿namespace NServiceBus.Persistence.AzureStorage.ComponentTests.Subscriptions
+{
+    class TestMessagea
+    {
+    }
+}
